add /create scoreboard-reference - inlined generateScorelines
